move from_name method from TestResource to NameConstructors
